PI-1153 added endpoints for actions

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/sentenceplan/service/ActionService.kt
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.sentenceplan.entity.ActionEntity
 import uk.gov.justice.digital.hmpps.sentenceplan.entity.ActionRepository
 import uk.gov.justice.digital.hmpps.sentenceplan.entity.ObjectiveRepository
-import uk.gov.justice.digital.hmpps.sentenceplan.entity.getBySentencePlanIdAndId
 import uk.gov.justice.digital.hmpps.sentenceplan.exception.NotFoundException
 import uk.gov.justice.digital.hmpps.sentenceplan.model.Action
 import uk.gov.justice.digital.hmpps.sentenceplan.model.ActionList
@@ -19,14 +18,13 @@ class ActionService(
   private val actionRepository: ActionRepository,
 ) {
   fun createAction(sentencePlanId: UUID, objectiveId: UUID, action: CreateAction): Action {
-    objectiveRepository.getBySentencePlanIdAndId(sentencePlanId, objectiveId)
+    if (!objectiveRepository.existsById(objectiveId)) throw NotFoundException("Objective", "id", objectiveId)
     val actionEntity = ActionEntity(
       objectiveId,
       action.description,
       action.interventionParticipation,
-      action.nationalInterventionCode,
-      action.accreditedProgramme,
-      action.localIntervention,
+      action.interventionName,
+      action.interventionType,
       action.status,
       action.owner,
     )
@@ -36,29 +34,28 @@ class ActionService(
   }
 
   fun updateAction(sentencePlanId: UUID, objectiveId: UUID, id: UUID, action: Action): Action {
-    objectiveRepository.getBySentencePlanIdAndId(sentencePlanId, objectiveId)
+    if (!objectiveRepository.existsById(objectiveId)) throw NotFoundException("Objective", "id", objectiveId)
     val actionEntity = actionRepository.findByIdOrNull(id) ?: throw NotFoundException("action", "id", id)
     actionEntity.description = action.description
-    actionEntity.accreditedProgramme = action.accreditedProgramme
-    actionEntity.localIntervention = action.localIntervention
+    actionEntity.interventionName = action.interventionName
+    actionEntity.interventionType = action.interventionType
     actionEntity.status = action.status
-    actionEntity.nationalInterventionCode = action.nationalInterventionCode
     actionEntity.owner = action.owner
     return actionRepository.save(actionEntity).toModel()
   }
 
   fun deleteAction(sentencePlanId: UUID, objectiveId: UUID, id: UUID) {
-    objectiveRepository.getBySentencePlanIdAndId(sentencePlanId, objectiveId)
+    if (!objectiveRepository.existsById(objectiveId)) throw NotFoundException("Objective", "id", objectiveId)
     actionRepository.deleteById(id)
   }
 
   fun listActions(sentencePlanId: UUID, objectiveId: UUID): ActionList {
-    objectiveRepository.getBySentencePlanIdAndId(sentencePlanId, objectiveId)
+    if (!objectiveRepository.existsById(objectiveId)) throw NotFoundException("Objective", "id", objectiveId)
     return ActionList(actionRepository.findAllByObjectiveIdOrderByCreatedDateTimeAsc(objectiveId).map { it.toModel() })
   }
 
   fun findAction(sentencePlanId: UUID, objectiveId: UUID, id: UUID): Action {
-    objectiveRepository.getBySentencePlanIdAndId(sentencePlanId, objectiveId)
+    if (!objectiveRepository.existsById(objectiveId)) throw NotFoundException("Objective", "id", objectiveId)
     return actionRepository.findByIdOrNull(id)?.toModel()
       ?: throw NotFoundException("action", "id", id)
   }
